Added template framework for applying transforms to distance maps
nnUNet currently does not have a framework for loading in distance maps that were calculated offline, though I have a version of nnUNet with a modified dataloader that can do this. For full functionality, distance maps should also be cropped and transformed in a similar fashion to the image and segmentation data. The easiest way to do this is to add another key in the data dictionary called 'dist_map'. If the data has this key, then a transform can call _apply_to_dist_map. Modifying the BasicTransform class is the simplest way to add this functionality to other transforms.